requirements.txt
flask==2.3.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
python-dotenv==1.0.0
openai==1.3.5
gunicorn==21.2.0
requests==2.31.0
flask-session==0.5.0 